Fix Xero OAuth state parameter not being passed to callback
The issue was that the state parameter was not being included in the XeroClient configuration.
The xero-node library requires the state to be passed in the client constructor options
so it can be included in the OAuth URL and returned in the callback.

This fixes the 'hasState: false' error seen in the Netlify function logs.